refactor test helpers
